refactor : 머지하면서 생긴 에러 수정
