+++ admin/src/components/BtnLight.js
@@ -1,6 +1,6 @@
 import classNames from "classnames";
 
-export function BtnLight({ text, onClick }) {
+export default function BtnLight({ text, onClick }) {
   return (
     <div className={classNames("btn-light")} onClick={onClick}>
       <span>{text}</span>
