+++ admin/src/pages/ReservationDetail.js
@@ -1,7 +1,7 @@
 import classNames from "classnames";
 import SideBar from "../components/SideBar";
-import { BtnDark } from "../components/BtnDark";
-import { BtnLight } from "../components/BtnLight";
+import BtnDark from "../components/BtnDark";
+import BtnLight from "../components/BtnLight";
 import { Link, useParams } from "react-router-dom";
 import { useEffect, useRef, useState } from "react";
 import AddImageModal from "../components/modal/AddImageModal";
