+++ admin/src/styles/styles.scss
@@ -5,7 +5,6 @@
 @import './pages/reservationList.scss';
 @import './components/button.scss';
 @import './pages/reservationList.scss';
-@import './pages/reservationDetail.scss';
 @import './components/infoTable.scss';
 @import './components/modal.scss';
